Add new infiniteScrollReset parameter to allow for easier list updates
This parameter is needed if you want to reuse the same infinite scroll directive and your items will completely change. A use case is a file browser where you navigate the folder structure and you want all the levels of the hierarchy to benefit from infinite scroll. Without a reset such as this, as soon as you scroll to the bottom of the list, subsequent lists that are not bigger than that will no longer have infinite scroll.

You can call the reset by adding an event listener and setting the reset to true. e.g. <div infinite-scroll="addMoreItems()" infinite-scroll-listen-for-event="list:changed" infinite-scroll-reset="true">
  [...]
</div>